requirements.txt
Django>=3.0.6, <3.1.0
djangorestframework>=3.11.0, <3.12.0
# psycopg2>=2.7.5, <2.8.0
psycopg2-binary==2.8.5
flake8>=3.8.0, <3.9.0
Pillow>=6.2.2

